backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
beautifulsoup4
tldextract
python-dotenv
openai